requirements.txt
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
